Backend/requirements.txt
asn1crypto==0.24.0
certifi==2018.1.18
cffi==1.12.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==2.5
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.0.2
django-cors-headers==2.2.0
django-crispy-forms==1.7.0
django-filter==1.1.0
django-formtools==2.1
django-guardian==1.4.9
django-import-export==0.7.0
django-redis==4.10.0
django-redis-cache==2.0.0
django-reversion==2.0.13
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
drf-extensions==0.4.0
et-xmlfile==1.0.1
future==0.16.0
httplib2==0.9.2
idna==2.6
itypes==1.1.0
jdcal==1.3
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
mysqlclient==1.3.12
Naked==0.1.31
oauthlib==3.0.1
odfpy==1.3.6
openpyxl==2.5.0
Pillow==5.0.0
pycparser==2.19
PyJWT==1.6.0
python3-openid==3.1.0
pytz==2018.3
PyYAML==3.12
raven==6.10.0
redis==3.2.0
requests==2.18.4
requests-oauthlib==1.2.0
shellescape==3.4.1
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.1.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
